scripts/requirements.txt
pypdf==5.4.0
ebooklib==0.18
beautifulsoup4==4.13.4
deep-translator==1.11.4
langdetect==1.0.9
Pillow==11.2.1
pytesseract==0.3.13
PyMuPDF==1.25.5
